restructure client calls & Makefile
