+++ pkg/client/http.go
@@ -1,4 +1,4 @@
-package utils
+package client
 
 import (
 	"bytes"
